requirements.txt
jupyter
jupyterthemes
chardet
requests

pyyaml
pytest
pytest-cov
fabric3